requirements.txt
requests
dotenv
python-docx
pandas
PyMuPDF
opencv-contrib-python
paddlepaddle
paddleocr
tqdm
google-generativeai
opencv-python
paddleocr
setuptools